+++ react/components/pages/Details/Shipping/ShippingCard.tsx
@@ -6,6 +6,7 @@ import { CSS, BASIC_CARD_WRAPPER } from '../../../../constants'
 import LabeledInfo from '../../../commons/LabeledInfo'
 import EditButton from '../../../commons/EditButton'
 import EditAlert from '../../../commons/EditAlert'
+import { SubscriptionsGroup } from '../'
 
 const messages = defineMessages({
   label: {
@@ -18,10 +19,7 @@ const messages = defineMessages({
   },
 })
 
-const ShippingCard: FunctionComponent<Props> = ({
-  onEdit,
-  subscriptionsGroup,
-}) => (
+const ShippingCard: FunctionComponent<Props> = ({ onEdit, group }) => (
   <div className={`${BASIC_CARD_WRAPPER} ${CSS.cardHorizontalPadding}`}>
     <div className="flex flex-row">
       <div className="db-s di-ns b f4 tl c-on-base">
@@ -29,27 +27,27 @@ const ShippingCard: FunctionComponent<Props> = ({
       </div>
       <div className="ml-auto">
         <EditButton
-          subscriptionStatus={subscriptionsGroup.status}
+          subscriptionStatus={group.status}
           onEdit={onEdit}
           testId="edit-address-button"
         />
       </div>
     </div>
     <div className="flex pt3-s pt5-ns w-100">
-      {subscriptionsGroup.shippingAddress ? (
+      {group.shippingAddress ? (
         <div className="w-100">
           <LabeledInfo labelId={messages.label.id}>
             <AddressRules
-              country={subscriptionsGroup.shippingAddress.country}
+              country={group.shippingAddress.country}
               shouldUseIOFetching
             >
-              <AddressSummary address={subscriptionsGroup.shippingAddress} />
+              <AddressSummary address={group.shippingAddress} />
             </AddressRules>
           </LabeledInfo>
         </div>
       ) : (
         <EditAlert
-          subscriptionStatus={subscriptionsGroup.status}
+          subscriptionStatus={group.status}
           onAction={onEdit}
           actionLabelMessage={messages.label}
           noActionMessage={messages.noAction}
@@ -62,7 +60,7 @@ const ShippingCard: FunctionComponent<Props> = ({
 )
 
 interface Props {
-  subscriptionsGroup: SubscriptionsGroupItemType
+  group: SubscriptionsGroup
   onEdit: () => void
 }