+++ react/components/pages/List/ItemDate.tsx
@@ -1,33 +1,40 @@
 import React, { FunctionComponent } from 'react'
 import { InjectedIntlProps, injectIntl } from 'react-intl'
 
-import { SubscriptionStatusEnum } from '../../../constants'
-
-interface Props {
-  item: SubscriptionsGroupItemType
-}
+import { SubscriptionStatus } from '../../../constants'
 
 const SubscriptionsGroupItemDate: FunctionComponent<Props &
-  InjectedIntlProps> = ({ item, intl }) => {
+  InjectedIntlProps> = ({
+  status,
+  nextPurchaseDate,
+  lastStatusUpdate,
+  intl,
+}) => {
   const content =
-    item.status === SubscriptionStatusEnum.Active
+    status === SubscriptionStatus.Active
       ? intl.formatMessage(
           { id: 'subscription.list.item.date.next.purchase' },
           {
-            date: intl.formatDate(item.nextPurchaseDate),
+            date: intl.formatDate(nextPurchaseDate),
           }
         )
       : intl.formatMessage(
           { id: 'subscription.list.item.date.since' },
           {
-            date: intl.formatDate(item.lastStatusUpdate),
+            date: intl.formatDate(lastStatusUpdate),
             status: intl.formatMessage({
-              id: `subscription.status.${item.status.toLowerCase()}`,
+              id: `subscription.status.${status.toLowerCase()}`,
             }),
           }
         )
 
   return <span className="t-small c-muted-2">{content}</span>
+}
+
+interface Props {
+  status: SubscriptionStatus
+  nextPurchaseDate: string
+  lastStatusUpdate: string
 }
 
 export default injectIntl(SubscriptionsGroupItemDate)