+++ react/components/pages/List/LoadingState.tsx
@@ -2,7 +2,6 @@ import React, { FunctionComponent } from 'react'
 
 import { CSS } from '../../../constants'
 import SkeletonLoader from '../../commons/SkeletonLoader'
-import DescriptionColumn from './SkeletonDescriptionColumn'
 
 const SubscriptionsGroupsLoading: FunctionComponent = () => {
   return (
@@ -18,7 +17,12 @@ const SubscriptionsGroupsLoading: FunctionComponent = () => {
             </div>
             <div className="mt3 w-70-ns w-100-s flex flex-row-s flex-column flex-nowrap flex-wrap-s">
               <div className="w-50-ns w-100-s">
-                <DescriptionColumn />
+                <div className="mt9-ns mt4-s">
+                  <SkeletonLoader width="70" size="3" />
+                  <div className="mt4">
+                    <SkeletonLoader width="60" size="3" />
+                  </div>
+                </div>
               </div>
             </div>
           </div>