+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [1.10.0] - 2019-12-12
+
+### Changed
+
+- Migrate to `subscriptions-graphql`.
+
 ## [1.9.1] - 2019-11-13
 
 ### Fixed